aruha-400 refactoring schema diff
Split the huge schema diff class into several smaller classes.
