+++ src/main/java/org/zalando/nakadi/validation/SchemaEvolutionService.java
@@ -14,7 +14,7 @@ import org.zalando.nakadi.domain.SchemaChange;
 import org.zalando.nakadi.domain.Version;
 import org.zalando.nakadi.exceptions.InvalidEventTypeException;
 import org.zalando.nakadi.validation.schema.ForbiddenAttributeIncompatibility;
-import org.zalando.nakadi.validation.schema.SchemaDiff;
+import org.zalando.nakadi.validation.schema.diff.SchemaDiff;
 import org.zalando.nakadi.validation.schema.SchemaEvolutionConstraint;
 import org.zalando.nakadi.validation.schema.SchemaEvolutionIncompatibility;
